S26: Implement trailing comments for Parameters
This implementation is a little different from the others because
the actual parameter may not be created until after the trailing
comments have been parsed (because the latter is parsed as whitespace).
So we have to set up a "dummy" parameter and transfer the trailing
comments after the fact.  I also didn't use an anonymous role to
store the trailing docs, because it was easier to extract them from
the dummy, should they exist at all.